[qa] Design Fix
- Review 및 Post 보완
