+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/recent/recent_trade/MyRecentTradeScreen.kt
@@ -2,16 +2,24 @@ package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.rece
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -24,9 +32,13 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import androidx.navigation.NavController
@@ -38,26 +50,38 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.plus
 import kotlinx.datetime.LocalDateTime
 import kr.linkerbell.campusmarket.android.common.util.coroutine.event.MutableEventFlow
+import kr.linkerbell.campusmarket.android.common.util.coroutine.event.eventObserve
 import kr.linkerbell.campusmarket.android.domain.model.feature.mypage.RecentTrade
 import kr.linkerbell.campusmarket.android.presentation.R
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Black
+import kr.linkerbell.campusmarket.android.presentation.common.theme.Body2
+import kr.linkerbell.campusmarket.android.presentation.common.theme.Caption1
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Caption2
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray600
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray900
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Headline1
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Headline2
+import kr.linkerbell.campusmarket.android.presentation.common.theme.Red400
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Space56
 import kr.linkerbell.campusmarket.android.presentation.common.theme.White
+import kr.linkerbell.campusmarket.android.presentation.common.util.compose.LaunchedEffectWithLifecycle
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.isEmpty
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.makeRoute
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.safeNavigate
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.safeNavigateUp
+import kr.linkerbell.campusmarket.android.presentation.common.view.DialogScreen
 import kr.linkerbell.campusmarket.android.presentation.common.view.RippleBox
+import kr.linkerbell.campusmarket.android.presentation.common.view.confirm.ConfirmButton
+import kr.linkerbell.campusmarket.android.presentation.common.view.confirm.ConfirmButtonProperties
+import kr.linkerbell.campusmarket.android.presentation.common.view.confirm.ConfirmButtonSize
+import kr.linkerbell.campusmarket.android.presentation.common.view.confirm.ConfirmButtonType
+import kr.linkerbell.campusmarket.android.presentation.common.view.textfield.TypingTextField
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.MyRecentTrade.my_recent_trade.MyRecentTradeArgument
+import kr.linkerbell.campusmarket.android.presentation.ui.main.home.MyRecentTrade.my_recent_trade.MyRecentTradeEvent
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.MyRecentTrade.my_recent_trade.MyRecentTradeIntent
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.MyRecentTrade.my_recent_trade.MyRecentTradeState
+import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.common.StarRatingBar
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.common.TradeHistoryCard
-import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.rating.RatingConstant
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.trade.info.TradeInfoConstant
 
 @Composable
@@ -74,6 +98,44 @@ fun MyRecentTradeScreen(
     val recentSalesTradeList by remember { mutableStateOf(data.recentSellTrades) }
 
     var selectedTab by remember { mutableIntStateOf(0) }
+
+    var isReviewDialogVisible by remember { mutableStateOf(false) }
+    var isReviewRequested by remember { mutableStateOf(false) }
+    var isReviewSuccessDialogVisible by remember { mutableStateOf(false) }
+
+    var targetUserId by remember { mutableStateOf(-1L) }
+    var itemId by remember { mutableStateOf(-1L) }
+
+    if (isReviewDialogVisible) {
+        ReviewDialog(
+            isReviewRequested = isReviewRequested,
+            onReviewButtonClicked = { userDescription, userRating ->
+                isReviewRequested = true
+                argument.intent(
+                    MyRecentTradeIntent.RateUser(
+                        targetUserId,
+                        itemId,
+                        userDescription,
+                        userRating
+                    )
+                )
+            },
+            onDismissRequest = {
+                isReviewDialogVisible = false
+            }
+        )
+    }
+
+    if (isReviewSuccessDialogVisible) {
+        DialogScreen(
+            title = "리뷰가 등록되었습니다!",
+            isCancelable = false,
+            onConfirm = { },
+            onDismissRequest = {
+                isReviewSuccessDialogVisible = false
+            }
+        )
+    }
 
     ConstraintLayout(
         modifier = Modifier
@@ -194,17 +256,29 @@ fun MyRecentTradeScreen(
                             )
                             navController.safeNavigate(tradeInfoRoute)
                         },
-                        onAddReviewClicked = { targetUserId, itemId ->
-                            val reviewInfoRoute = makeRoute(
-                                route = TradeInfoConstant.ROUTE,
-                                arguments = mapOf(
-                                    RatingConstant.ROUTE_ARGUMENT_USER_ID to targetUserId,
-                                    RatingConstant.ROUTE_ARGUMENT_ITEM_ID to itemId
-                                )
-                            )
-                            navController.safeNavigate(reviewInfoRoute)
+                        onAddReviewClicked = {
+                            targetUserId = if (trade.userId == data.myId) trade.buyerId
+                            else data.myId
+                            itemId = trade.itemId
+                            isReviewDialogVisible = true
                         }
                     )
+                }
+            }
+        }
+    }
+
+    LaunchedEffectWithLifecycle(event, coroutineContext) {
+        event.eventObserve { event ->
+            when (event) {
+                is MyRecentTradeEvent.RateSuccess -> {
+                    isReviewSuccessDialogVisible = true
+                    isReviewDialogVisible = false
+                    argument.intent(MyRecentTradeIntent.RefreshTradeList)
+                }
+
+                is MyRecentTradeEvent.RateFail -> {
+                    isReviewRequested = false
                 }
             }
         }
@@ -241,6 +315,121 @@ private fun MyRecentTradeListTab(
                 color = Gray900,
                 modifier = Modifier.padding(horizontal = 4.dp)
             )
+        }
+    }
+}
+
+@Composable
+private fun ReviewDialog(
+    onReviewButtonClicked: (String, Int) -> Unit,
+    onDismissRequest: () -> Unit,
+    isReviewRequested: Boolean
+) {
+    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
+    val screenHeight = LocalConfiguration.current.screenHeightDp.dp
+
+    var userRating by remember { mutableIntStateOf(5) }
+    var userDescription by remember { mutableStateOf("") }
+
+    var descriptionLength by remember { mutableIntStateOf(0) }
+
+    Dialog(
+        onDismissRequest = {},
+    ) {
+        Column(
+            modifier = Modifier
+                .size(height = screenHeight * 0.6f, width = screenWidth * 0.8f)
+                .clip(RoundedCornerShape(16.dp))
+                .fillMaxSize()
+                .background(Color.White)
+                .verticalScroll(rememberScrollState()),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.SpaceBetween
+        ) {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                Text(
+                    text = "거래는 어떠셨나요?",
+                    style = Headline2,
+                    modifier = Modifier.padding(8.dp)
+                )
+                Text(
+                    text = "상대방에 대한 리뷰를 남겨주세요",
+                    style = Body2,
+                    modifier = Modifier.padding(vertical = 8.dp)
+                )
+                Text(
+                    text = userRating.toString(),
+                    style = Headline1,
+                    modifier = Modifier.padding(8.dp)
+                )
+                StarRatingBar(
+                    rating = userRating,
+                    180.dp,
+                    onRatingChanged = { newRating ->
+                        userRating = newRating
+                    }
+                )
+
+                TypingTextField(
+                    text = userDescription,
+                    onValueChange = {
+                        if (descriptionLength <= 200) {
+                            userDescription = it
+                            descriptionLength = userDescription.length
+                        }
+                    },
+                    maxLines = 100,
+                    modifier = Modifier
+                        .heightIn(min = 140.dp)
+                        .padding(horizontal = 16.dp, vertical = 8.dp)
+                )
+                Text(
+                    text = if (descriptionLength <= 200) "(${descriptionLength}/200)"
+                    else "리뷰는 최대 200자까지만 작성할 수 있어요!",
+                    modifier = Modifier
+                        .align(Alignment.End)
+                        .padding(horizontal = 16.dp),
+                    style = Caption1,
+                    color = if (descriptionLength <= 200) Gray900 else Red400
+                )
+            }
+            Column(modifier = Modifier.padding(16.dp)) {
+                Row(modifier = Modifier.wrapContentSize()) {
+                    ConfirmButton(
+                        properties = ConfirmButtonProperties(
+                            size = ConfirmButtonSize.Large,
+                            type = ConfirmButtonType.Secondary
+                        ),
+                        modifier = Modifier.weight(1f),
+                        onClick = {
+                            onDismissRequest()
+                        }
+                    ) { style ->
+                        Text(
+                            text = "나중에 할게요",
+                            style = style
+                        )
+                    }
+                    Spacer(modifier = Modifier.width(10.dp))
+                    ConfirmButton(
+                        properties = ConfirmButtonProperties(
+                            size = ConfirmButtonSize.Large,
+                            type = ConfirmButtonType.Primary
+                        ),
+                        modifier = Modifier.weight(1f),
+                        onClick = {
+                            if (!isReviewRequested) {
+                                onReviewButtonClicked(userDescription, userRating)
+                            }
+                        }
+                    ) { style ->
+                        Text(
+                            text = "평가하기",
+                            style = style
+                        )
+                    }
+                }
+            }
         }
     }
 }
