+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/recent/recent_trade/MyRecentTradeViewModel.kt
@@ -6,6 +6,8 @@ import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -16,6 +18,7 @@ import kr.linkerbell.campusmarket.android.common.util.coroutine.event.asEventFlo
 import kr.linkerbell.campusmarket.android.domain.model.feature.mypage.RecentTrade
 import kr.linkerbell.campusmarket.android.domain.model.nonfeature.error.ServerException
 import kr.linkerbell.campusmarket.android.domain.usecase.feature.myprofile.GetRecentTradeListUseCase
+import kr.linkerbell.campusmarket.android.domain.usecase.feature.trade.RateUserUseCase
 import kr.linkerbell.campusmarket.android.presentation.common.base.BaseViewModel
 import kr.linkerbell.campusmarket.android.presentation.common.base.ErrorEvent
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.MyRecentTrade.my_recent_trade.MyRecentTradeEvent
@@ -26,7 +29,8 @@ import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.other
 @HiltViewModel
 class MyRecentTradeViewModel @Inject constructor(
     private val savedStateHandle: SavedStateHandle,
-    private val getRecentTradeListUseCase: GetRecentTradeListUseCase
+    private val getRecentTradeListUseCase: GetRecentTradeListUseCase,
+    private val rateUserUseCase: RateUserUseCase
 ) : BaseViewModel() {
 
     private val _state: MutableStateFlow<MyRecentTradeState> =
@@ -81,7 +85,35 @@ class MyRecentTradeViewModel @Inject constructor(
                     getBuyTradeHistory(_userId.value)
                 }
             }
+
+            is MyRecentTradeIntent.RateUser -> {
+                launch {
+                    rateUser(
+                        intent.targetUserId,
+                        intent.itemId,
+                        intent.description,
+                        intent.rating
+                    )
+                }
+            }
+
+            MyRecentTradeIntent.RefreshTradeList -> {
+                launch {
+                    refreshAllHistory(_userId.value)
+                }
+            }
         }
+    }
+
+    private suspend fun refreshAllHistory(userId: Long) = coroutineScope {
+
+        val getSalesTradeHistoryJob = async { getSalesTradeHistory(userId) }
+        val getBuyTradeHistoryJob = async { getBuyTradeHistory(userId) }
+        val getAllTradeHistoryJob = async { getAllTradeHistory(userId) }
+
+        getSalesTradeHistoryJob.await()
+        getBuyTradeHistoryJob.await()
+        getAllTradeHistoryJob.await()
     }
 
     private suspend fun getSalesTradeHistory(userId: Long) {
@@ -136,5 +168,33 @@ class MyRecentTradeViewModel @Inject constructor(
             }.collect {
                 _recentTrades.value = it
             }
+    }
+
+    private suspend fun rateUser(
+        targetUserId: Long,
+        itemId: Long,
+        description: String,
+        rating: Int
+    ) {
+        rateUserUseCase(
+            targetUserId = targetUserId,
+            itemId = itemId,
+            description = description,
+            rating = rating
+        ).onSuccess {
+            _state.value = MyRecentTradeState.Init
+            _event.emit(MyRecentTradeEvent.RateSuccess)
+        }.onFailure { exception ->
+            when (exception) {
+                is ServerException -> {
+                    _errorEvent.emit(ErrorEvent.InvalidRequest(exception))
+                }
+
+                else -> {
+                    _errorEvent.emit(ErrorEvent.UnavailableServer(exception))
+                }
+            }
+            _event.emit(MyRecentTradeEvent.RateFail)
+        }
     }
 }
